FINALLY IMPLEMENTED FUNCTIONALITY to the DaysUntilFragment
Took me forever to find a formula
Thought I could just extract out years,months,weeks,day by just using a millisecond difference,
but it turned out to be WAAAY more complicated than that.
Mainly because I overlooked the facts that: 1)Months don't always have the same # of days, and the order at which
they don't exactly switch between 30 and 31 days each time.
2) 1 month != 4 weeks. In fact, creating this whole functionality made me question the notion of
why we even use the term "month", like we denote 4 weeks to a month or 30 - 31 days but that's not
always correct or useful, especially if you're starting from a day in the middle of the month
(ex: 9/5 - 10/5 is 30 days, but 10/5 - 11/5 is 31 days). It turns out, the # of days for that
is determined by the # of days in the starting month.

I know what you're thinking, pretty common sense stuff, right? Oh well, it was interesting to kind of
look into this anyways.

2. UnitFragment:
   Replaced keyboard with keypad by adding in the android:inputType="number" attribute to the elements that needed it

3. Strings.xml
   Set the resulttext and resultdetailstext to be null initially

Things that still need to be done:
1. Implement onSaveInstance for DaysUntilFragment
2. Fix the arrayoutofbounds exception that occurs in UnitFragment when you press too many buttons
3. GET THE UNITFRAGMENT TO SHOW UP INITIALLY (????)